rename shader module
