+++ frontend/javascripts/oxalis/geometries/materials/node_shader.ts
@@ -1,7 +1,7 @@
 import * as THREE from "three";
 import { ViewModeValues, ViewModeValuesIndices } from "oxalis/constants";
 import type { Uniforms } from "oxalis/geometries/materials/plane_material_factory";
-import { formatNumberAsGLSLFloat } from "oxalis/shaders/main_data_fragment.glsl";
+import { formatNumberAsGLSLFloat } from "oxalis/shaders/main_data_shaders.glsl";
 import { getBaseVoxel } from "oxalis/model/scaleinfo";
 import { getZoomValue } from "oxalis/model/accessors/flycam_accessor";
 import { listenToStoreProperty } from "oxalis/model/helpers/listener_helpers";
