+++ frontend/javascripts/oxalis/geometries/materials/plane_material_factory.ts
@@ -46,7 +46,7 @@ import type { DatasetLayerConfiguration } from "oxalis/store";
 import Store from "oxalis/store";
 import * as Utils from "libs/utils";
 import app from "app";
-import getMainFragmentShader, { getMainVertexShader } from "oxalis/shaders/main_data_fragment.glsl";
+import getMainFragmentShader, { getMainVertexShader } from "oxalis/shaders/main_data_shaders.glsl";
 import shaderEditor from "oxalis/model/helpers/shader_editor";
 import type { ElementClass } from "types/api_flow_types";
 import { CuckooTable } from "oxalis/model/bucket_data_handling/cuckoo_table";
